+++ mypyc/genfunc.py
@@ -29,7 +29,7 @@ from mypyc.common import (
 from mypyc.sametype import is_same_method_signature
 from mypyc.genopsutil import concrete_arg_kind, is_constant, add_self_to_env
 from mypyc.genopscontext import FuncInfo, GeneratorClass, ImplicitClass
-from mypyc.genstatement import transform_try_except
+from mypyc.irbuild.statement import transform_try_except
 from mypyc.genops import IRBuilder
